Move ichnaea.async to ichnaea.taskapp
async is a reserved keyword in Python 3.5, so it shouldn't be used as a
module name.

Change the name from "async" to "taskapp", to mirror the webapp module.
Adjust the docs to talk about task workers rather than async workers.
