+++ docker/run_scheduler.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
 # BBB: Celery 4
-# exec /app/bin/celery -A ichnaea.async.app:celery_app beat --loglevel='WARNING'
+# exec /app/bin/celery -A ichnaea.taskapp.app:celery_app beat --loglevel='WARNING'
 
-celery -A ichnaea.async.app:celery_app beat \
+celery -A ichnaea.taskapp.app:celery_app beat \
     -s "/var/run/location/celerybeat-schedule" \
     --pidfile="/var/run/location/celerybeat.pid"
