+++ ichnaea/taskapp/app.py
@@ -5,7 +5,7 @@ from celery import Celery
 from celery.app import app_or_default
 from celery.signals import beat_init, worker_process_init, worker_process_shutdown
 
-from ichnaea.async.config import (
+from ichnaea.taskapp.config import (
     configure_celery,
     init_beat,
     init_worker,
@@ -18,7 +18,7 @@ def init_beat_process(signal, sender, **kw):
     """
     Called automatically when `celery beat` is started.
 
-    Calls :func:`ichnaea.async.config.init_beat`.
+    Calls :func:`ichnaea.taskapp.config.init_beat`.
     """
     celery_app = app_or_default()
     init_beat(sender, celery_app)
@@ -30,7 +30,7 @@ def init_worker_process(signal, sender, **kw):
     Called automatically when `celery worker` is started. This is executed
     inside each forked worker process.
 
-    Calls :func:`ichnaea.async.config.init_worker`.
+    Calls :func:`ichnaea.taskapp.config.init_worker`.
     """
     # get the app in the current worker process
     celery_app = app_or_default()
@@ -43,12 +43,12 @@ def shutdown_worker_process(signal, sender, **kw):
     Called automatically when `celery worker` is stopped. This is executed
     inside each forked worker process.
 
-    Calls :func:`ichnaea.async.config.shutdown_worker`.
+    Calls :func:`ichnaea.taskapp.config.shutdown_worker`.
     """
     celery_app = app_or_default()
     shutdown_worker(celery_app)
 
 
-celery_app = Celery("ichnaea.async.app")
+celery_app = Celery("ichnaea.taskapp.app")
 
 configure_celery(celery_app)
